use registry cache and only download digest artifacts
Since I last looked at this CI in detail a new cache backend has been included in docker's buildkit called a registry cache. Basically, the builder pushes all of the intermediate layers to a specific image tag on a registry so that later builders can pull those layers if their hash matches.

https://docs.docker.com/build/cache/backends/registry/

I am unsure if we will run into limitations by DockerHub for this choice, but it hopefully will increase the number of cache hits and therefore make building faster. As far as I can tell DockerHub only has limits on the number of push/pulls and not the overall size of the images.